huge refactor
